requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.12.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
